+++ Models/IClientTeamRepository.cs
@@ -3,7 +3,7 @@
     public interface IClientTeamRepository
     {
         public IEnumerable<ClientTeam> AllTeams { get; }
-        public ClientTeam? GetTeamById(Int16? id);
+        public ClientTeam? GetTeamById(int? id);
         public ClientTeam? GetTeamByName(string teamName);
         public IEnumerable<ClientTeam> GetTeamsByFinancialContact(string contact);
         public IEnumerable<ClientTeam> GetTeamsByPrimaryContact(string primaryContact);
